classifier/requirements.txt
scikit-learn==0.22
scipy==1.7.2
propy3 ==1.0.2
numpy ==1.21.4
jupyter ==1.0.0
isoelectric==1.0
peptides==0.3.2
pandas==1.3.5
